バグ修正

+++ sql/CreateTable.sql
@@ -13,9 +13,13 @@ CREATE TABLE BTC_VALUE(
 -- ユーザ情報
 -- DROP TABLE USERDATA;
 CREATE TABLE USERDATA(
-	Money bigint,				-- 所持金額
-	BitCoin bigint,				-- 所持ビットコイン量
-	BitCoinBuyMoney bigint,		-- 取得時現金
-	BitCoinBuyValue bigint,		-- 取得時ビットコイン価格
+	Money numeric(15,8),				-- 所持金額
+	BitCoin numeric(15,8),				-- 所持ビットコイン量
+	BitCoinBuyMoney numeric(15,8),		-- 取得時現金
+	BitCoinBuyValue numeric(15,8),		-- 取得時ビットコイン価格
 	BitCoinFlag Boolean			-- ビットコイン所持フラグ
 );
+-- テストデータ
+-- insert into userdata values (10000,0,0,0,FALSE);
+
+
